requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
astor==0.8.1
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.21
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1596748257051/work
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
Django==3.0.3
docutils==0.16
entrypoints==0.3
flake8==3.8.3
future==0.18.2
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1596863485713/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
intervaltree @ file:///tmp/build/80754af9/intervaltree_1594361675072/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1596868620883/work
ipython-genutils==0.2.0
isort @ file:///C:/ci/isort_1597318826467/work
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
keras-resnet==0.2.0
keyring @ file:///C:/ci/keyring_1596867296032/work
kiwisolver==1.2.0
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1597440772084/work
Markdown @ file:///C:/ci/markdown_1597415185889/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-base_1592846084747/work
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
networkx @ file:///tmp/build/80754af9/networkx_1594377231366/work
notebook @ file:///C:/ci/notebook_1596837179121/work
numpy==1.19.1
numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.40
opt-einsum==3.1.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
pexpect @ file:///C:/ci/pexpect_1594383465186/work
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
pluggy==0.13.1
progressbar2==3.51.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1592848020240/work
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
PyNaCl @ file:///C:/ci/pynacl_1595009241355/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1594397536060/work
python-language-server @ file:///C:/ci/python-language-server_1594152769556/work
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1594392691209/work
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
Rtree==0.9.4
scikit-image==0.16.2
scipy @ file:///C:/ci/scipy_1592916958183/work
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx @ file:///tmp/build/80754af9/sphinx_1597428793432/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder @ file:///C:/ci/spyder_1594826266791/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1597006281278/work
sqlparse==0.3.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
ujson==1.35
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
watchdog @ file:///C:/ci/watchdog_1593446515400/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zipp==3.1.0
